+++ src/braket/default_simulator/simulator.py
@@ -22,7 +22,10 @@ import numpy as np
 from braket.default_simulator.observables import Hermitian, TensorProduct
 from braket.default_simulator.openqasm.circuit import Circuit
 from braket.default_simulator.openqasm.interpreter import Interpreter
-from braket.default_simulator.openqasm.program_context import AbstractProgramContext, ProgramContext
+from braket.default_simulator.openqasm.program_context import (
+    AbstractProgramContext,
+    ProgramContext,
+)
 from braket.default_simulator.operation import Observable, Operation
 from braket.default_simulator.operation_helpers import from_braket_instruction
 from braket.default_simulator.result_types import (
@@ -161,7 +164,7 @@ class BaseLocalSimulator(OpenQASMSimulator):
         return self.run_jaqcd(circuit_ir, *args, **kwargs)
 
     def create_program_context(self) -> AbstractProgramContext:
-        return ProgramContext()
+        return ProgramContext(simulator=self)
 
     @abstractmethod
     def initialize_simulation(self, **kwargs) -> Simulation: