+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v1.39.2 (2026-05-11)
+
+### Bug Fixes and Other Changes
+
+ * Rename `Measure` to `Projection`
+
+## v1.39.1 (2026-05-04)
+
+### Bug Fixes and Other Changes
+
+ * shallow-copy circuit instructions into initial branched path
+ * Delegate branch calculation to correct simulator
+
 ## v1.39.0 (2026-04-29)
 
 ### Features